% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebook.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Generate a Codebook for a Data Frame}
\usage{
generate_codebook(
  df,
  hide_statistics = NULL,
  top_n = 5,
  extra_vars = NULL,
  extra_key = NULL
)
}
\arguments{
\item{df}{A data frame for which to generate the codebook.}

\item{hide_statistics}{An optional character vector of column names for which statistics should be hidden. Useful for personally identifiable information.}

\item{top_n}{An integer specifying the number of top categories to display for categorical variables. Default is 5.}

\item{extra_vars}{An optional data frame to merge additional metadata, such as descriptions or field types.}

\item{extra_key}{A character specifying the column used for merging the \code{extra_vars} table.}
}
\value{
A formatted codebook as a \code{kable} object.
}
\description{
This function generates a codebook summarizing the variables in a data frame, including
statistics for numeric, categorical, logical, and date variables. The codebook can be output
in \code{kable} format, with support for merging additional metadata.
}
