% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook.R
\name{get_date_stats}
\alias{get_date_stats}
\title{Calculate Date Statistics}
\usage{
get_date_stats(var)
}
\arguments{
\item{var}{A Date vector.}
}
\value{
A formatted string with the minimum and maximum dates.
}
\description{
Helper function to calculate statistics for Date variables.
}
\keyword{internal}
