% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook.R
\name{get_logical_stats}
\alias{get_logical_stats}
\title{Calculate Logical Statistics}
\usage{
get_logical_stats(var)
}
\arguments{
\item{var}{A logical vector.}
}
\value{
A formatted string with counts and percentages of TRUE and FALSE values.
}
\description{
Helper function to calculate statistics for logical variables.
}
\keyword{internal}
