% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook.R
\name{get_missing_info}
\alias{get_missing_info}
\title{Calculate Missing Value Information}
\usage{
get_missing_info(var)
}
\arguments{
\item{var}{A vector of any type.}
}
\value{
A list containing the number and percentage of missing values.
}
\description{
Helper function to calculate missing value information.
}
\keyword{internal}
