% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebook.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Generate a Codebook}
\usage{
generate_codebook(data)
}
\arguments{
\item{data}{A data frame for which to generate the codebook.}
}
\value{
A data frame or list containing variable descriptions.
}
\description{
This function creates a codebook for a given dataset.
}
